refactor(nomt-core): bundle all proof types and function in a single module
